Start implementing multiple file handling
